se creo la funcionalidad para editar productos

+++ src/main/java/com/comercio/Service/ProductoService.java
@@ -13,9 +13,9 @@ public interface ProductoService {
  
     public List<Producto> listarProductos();
     public Producto save (Producto producto);
-    public Optional<Producto> get (Integer id);
+    public Optional<Producto> get (Long id);
     public void update (Producto producto);
-    public void delete(Integer id);
+    public void delete(long id);
     
     
 }
